MedReminder/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
celery>=5.2
django-celery-beat>=2.2
twilio>=8.0
chatterbot>=1.0.5
chatterbot-corpus>=1.2.0
WeasyPrint>=59.0
django-widget-tweaks>=1.4.12
python-dotenv>=1.0.0
requests>=2.28 